# openapi.yml
openapi: 3.1.0
info:
  description: The same-origin contract consumed by the BeyondPilot web application.
  title: BeyondPilot API
  version: 1.0.0
servers:
- description: Same origin as the web application
  url: /
...
paths:
...
  /api/organization/admin/claims/{id}/approve:
    post:
      operationId: approveOrganizationClaim
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The person owns the organization.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The claim is not open.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The person joined another organization in the meantime.
      security:
      - session: []
      summary: Let a person own an organization nobody owns
      tags:
      - Organization administration
  /api/organization/admin/claims/{id}/decline:
    post:
      operationId: declineOrganizationClaim
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The claim is closed.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The claim is not open.
      security:
      - session: []
      summary: Decline a request to own an organization
      tags:
      - Organization administration
  /api/organization/admin/organizations:
    get:
      operationId: listAdminOrganizations
      parameters:
      - description: "Organizations whose name or domain contains this, ignoring case."
        in: query
        name: q
        required: false
        schema:
          type:
          - string
          - "null"
          maxLength: 100
          minLength: 0
      - description: Only organizations of this review status.
        in: query
        name: status
        required: false
        schema:
          type: string
          enum:
          - pending
          - approved
          - rejected
          pattern: pending|approved|rejected
      - description: "The page, counted from 1."
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganizationList"
          description: One page of the organizations the parameters select.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A parameter is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
      security:
      - session: []
      summary: "The organizations, those waiting for review first"
      tags:
      - Organization administration
    post:
      operationId: createAdminOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateOrganization"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganization"
          description: "The organization, approved and without an owner."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Another organization has the email domain.
      security:
      - session: []
      summary: Create an organization for a company that is not here yet
      tags:
      - Organization administration
  /api/organization/admin/organizations/{id}:
    get:
      operationId: getAdminOrganization
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganization"
          description: "The organization, its people and the requests to own it."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such organization.
      security:
      - session: []
      summary: One organization as an operator reviews it
      tags:
      - Organization administration
  /api/organization/admin/organizations/{id}/approve:
    post:
      operationId: approveOrganization
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The organization is approved.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such organization.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The organization is not waiting for review.
      security:
      - session: []
      summary: Approve an organization that waits for review
      tags:
      - Organization administration
  /api/organization/admin/organizations/{id}/refuse:
    post:
      operationId: refuseOrganization
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefuseOrganization"
        required: true
      responses:
        "204":
          description: "The organization is refused, with the reason."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such organization.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The organization is not waiting for review.
      security:
      - session: []
      summary: Refuse an organization that waits for review
      tags:
      - Organization administration
...
  /api/organization/organizations:
    get:
      operationId: searchOrganizations
      parameters:
      - description: "A part of the name, ignoring case."
        in: query
        name: q
        required: false
        schema:
          type: string
          maxLength: 100
          minLength: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationSearch"
          description: "Ten organizations at most, by name; none for a text shorter\
            \ than two characters."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The text is too long.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
      security:
      - session: []
      summary: The approved organizations whose name contains a text
      tags:
      - Organizations
    post:
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganization"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
          description: "The organization, waiting for GenAI Fund's review."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller already belongs to an organization, or waits on\
            \ a request to join one."
      security:
      - session: []
      summary: Create an organization the caller owns
      tags:
      - Organizations
...
  /api/organization/organizations/{slug}:
    get:
      operationId: getOrganization
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrganization"
          description: The organization.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: No approved organization has this address.
      summary: "The public page of an approved organization, read without a session"
      tags:
      - Organizations
...
components:
  schemas:
...
    AdminCreateOrganization:
      type: object
      description: "An organization an operator creates for a company that is not\
        \ here yet. It is approved from the start and has no member until someone\
        \ accepts the invitation to own it, or claims it."
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
          pattern: "[A-Z]{2}"
        emailDomain:
          type:
          - string
          - "null"
          description: The domain of the company's work addresses. The first person
            who signs in on it may own the organization at once.
          example: tasco.com.vn
          maxLength: 253
          minLength: 0
          pattern: "[a-z0-9]([a-z0-9-]*[a-z0-9])?(\\.[a-z0-9]([a-z0-9-]*[a-z0-9])?)+"
        name:
          type: string
          maxLength: 120
          minLength: 0
        ownerEmail:
          type:
          - string
          - "null"
          format: email
          description: The address invited to own it.
          maxLength: 254
          minLength: 0
        roles:
          type: array
          items:
            type: string
            pattern: provider|enterprise
          maxItems: 2
          minItems: 1
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
          pattern: company|builder_team|independent_builder|other
        website:
          type:
          - string
          - "null"
          maxLength: 300
          minLength: 0
          pattern: "https?://[^\\s]+"
      required:
      - name
      - roles
      - type
    AdminOrganization:
      type: object
      description: One organization as an operator reviews it.
      properties:
        claims:
          type: array
          description: The open requests to own it; empty for an owned organization.
          items:
            $ref: "#/components/schemas/OrganizationJoinRequest"
        createdBy:
          type: string
          description: "Who created it, as they are shown."
        createdByEmail:
          type: string
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationInvitation"
        members:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMember"
        organization:
          $ref: "#/components/schemas/Organization"
      required:
      - claims
      - createdBy
      - createdByEmail
      - invitations
      - members
      - organization
    AdminOrganizationList:
      type: object
      description: "One page of organizations: those waiting for review first, then\
        \ the newest."
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrganizationSummary"
        page:
          type: integer
          format: int32
          description: "The page returned, counted from 1."
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
          description: "How many organizations match, over all pages."
      required:
      - items
      - page
      - pageSize
      - total
    AdminOrganizationSummary:
      type: object
      description: One organization in the operators' list.
      properties:
        country:
          type:
          - string
          - "null"
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        members:
          type: integer
          format: int32
        name:
          type: string
        openClaims:
          type: integer
          format: int32
          description: How many people ask to own it; always 0 for an owned organization.
        owned:
          type: boolean
          description: Whether a person owns it; an operator-created organization
            has no owner until someone accepts or claims it.
        roles:
          type: array
          items:
            type: string
        slug:
          type: string
        status:
          type: string
          enum:
          - pending
          - approved
          - rejected
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
      required:
      - createdAt
      - id
      - members
      - name
      - openClaims
      - owned
      - roles
      - slug
      - status
      - type
...
    CreateOrganization:
      type: object
      description: "A new organization. Its creator owns it, and it waits for GenAI\
        \ Fund's review."
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
          pattern: "[A-Z]{2}"
        description:
          type:
          - string
          - "null"
          maxLength: 2000
          minLength: 0
        name:
          type: string
          maxLength: 120
          minLength: 0
        roles:
          type: array
          description: One or both of `provider` and `enterprise`.
          items:
            type: string
            pattern: provider|enterprise
          maxItems: 2
          minItems: 1
        teamSize:
          type:
          - string
          - "null"
          pattern: just_me|2_9|10_49|50_99|100_499|500_999|1000_4999|5000_plus
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
          pattern: company|builder_team|independent_builder|other
        website:
          type:
          - string
          - "null"
          maxLength: 300
          minLength: 0
          pattern: "https?://[^\\s]+"
      required:
      - name
      - roles
      - type
...
    Organization:
      type: object
      description: "An organization as the people who belong to it, and operators,\
        \ see it."
      properties:
        autoJoin:
          type: boolean
          description: Whether an address on the domain joins at once; otherwise it
            asks the owners.
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        createdAt:
          type: string
          format: date-time
        decisionMessage:
          type:
          - string
          - "null"
          description: What the operator wrote to the owners with the refusal.
        decisionReason:
          type:
          - string
          - "null"
          description: Why it was last refused.
          enum:
          - duplicate
          - not_a_real_organization
          - incomplete
          - out_of_scope
          - other
        description:
          type:
          - string
          - "null"
        emailDomain:
          type:
          - string
          - "null"
          description: The domain whose addresses may join; null when it was made
            from a public mail address.
        id:
          type: string
          format: uuid
        name:
          type: string
        roles:
          type: array
          description: "What it does here: `provider` lists AI solutions, `enterprise`\
            \ posts use cases."
          items:
            type: string
        slug:
          type: string
        status:
          type: string
          description: GenAI Fund's review of the organization.
          enum:
          - pending
          - approved
          - rejected
        teamSize:
          type:
          - string
          - "null"
          description: Null is unknown.
          enum:
          - just_me
          - 2_9
          - 10_49
          - 50_99
          - 100_499
          - 500_999
          - 1000_4999
          - 5000_plus
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
        version:
          type: integer
          format: int64
          description: "Sent back with a save, which is refused when the organization\
            \ changed since."
        website:
          type:
          - string
          - "null"
      required:
      - autoJoin
      - createdAt
      - id
      - name
      - roles
      - slug
      - status
      - type
      - version
...
    OrganizationJoinRequest:
      type: object
      description: An open request to join an organization.
      properties:
        claim:
          type: boolean
          description: "Whether nobody owns the organization, so GenAI Fund decides\
            \ and approval makes the person its owner."
        createdAt:
          type: string
          format: date-time
        email:
          type: string
        id:
          type: string
          format: uuid
        message:
          type:
          - string
          - "null"
        name:
          type:
          - string
          - "null"
          description: The name of the person who asks; null until they have one.
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
      required:
      - claim
      - createdAt
      - email
      - id
      - organizationId
      - organizationName
...
    OrganizationSearch:
      type: object
      description: "The approved organizations a search finds, by name; ten at most."
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrganizationMatch"
      required:
      - items
    Problem:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          description: Stable failure code for module failures and request validation;
            clients branch on it.
          maxLength: 63
          pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
        detail:
          type: string
          description: Safe human-readable explanation; never branch on it.
        errors:
          type: array
          description: "Request validation only: one entry per violation. A rejected\
            \ value is never echoed."
          items:
            type: object
            additionalProperties: false
            properties:
              code:
                type: string
                description: "Constraint code, for example `NOT_BLANK` or `SIZE`."
                pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
              detail:
                type: string
                description: Safe fallback text; clients translate by code.
              params:
                type: object
                additionalProperties: false
                description: Numeric bounds of the constraint; omitted when it has
                  none.
                properties:
                  max:
                    type: number
                  min:
                    type: number
              pointer:
                type: string
                description: "JSON Pointer to the offending member, for example `#/title`."
            required:
            - code
            - detail
            - pointer
        instance:
          type: string
          format: uri-reference
          description: Request path that produced the problem.
        requestId:
          type: string
          format: uuid
          description: Identifier of the request in the server logs and the `X-Request-Id`
            header.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        title:
          type: string
          description: Short summary of the problem kind; never branch on it.
        type:
          type: string
          format: uri
          description: '`urn:beyondpilot:failure:<code in kebab case>` when `code`
            is present; omitted means `about:blank`.'
      required:
      - requestId
      - status
      - title
...
    PublicOrganization:
      type: object
      description: An approved organization as anyone reads it.
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        description:
          type:
          - string
          - "null"
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum:
          - company
          - builder_team
          - independent_builder
          - other
        website:
          type:
          - string
          - "null"
      required:
      - name
      - slug
      - type
...
    RefuseOrganization:
      type: object
      description: "Why an organization is not approved, and what its owners are told."
      properties:
        message:
          type:
          - string
          - "null"
          description: Shown to the owners with the refusal.
          maxLength: 1000
          minLength: 0
        reason:
          type: string
          enum:
          - duplicate
          - not_a_real_organization
          - incomplete
          - out_of_scope
          - other
          pattern: duplicate|not_a_real_organization|incomplete|out_of_scope|other
      required:
      - reason
...
  securitySchemes:
    session:
      description: "The session cookie set by signing in. Requests that change state\
        \ also send `X-BeyondPilot-CSRF: 1`."
      in: cookie
      name: BEYONDPILOT_SESSION
      type: apiKey